Retrieve initial JWKS when compute is empty
In the service, empty pre-warmed computes sit around in pools to assist
with cold starts. In the past, they received empty compute specs from
the control plane. With authorized communication from the control plane,
we need access to the JWKS prior to the HTTP server starting to setup
the middleware. Muhammet added support to the control plane to always
send back a JWKS to empty computes. Thus, we will never not have at
least a partial spec, which most of this patch accounts for.

Signed-off-by: Tristan Partin <tristan@neon.tech>